Fix login api route, and basic home and dashboard
