+++ frontend/src/app/page.tsx
@@ -1,25 +1,13 @@
-import Link from 'next/link';
-
-export default function Home() {
+import Header from '../components/Header';
+export default function Dashboard() {
   return (
     <div className="min-h-screen bg-gray-100">
-      <header className="bg-blue-600 text-white p-4">
-        <div className="container mx-auto flex justify-between">
-          <h1 className="text-xl font-bold">StackVarsity</h1>
-          <nav>
-            <Link href="/auth/login" className="mr-4">Login</Link>
-            <Link href="/auth/signup">Signup</Link>
-          </nav>
-        </div>
-      </header>
+      <Header />
       <main className="container mx-auto py-10">
-        <h2 className="text-4xl font-bold mb-6">Learn Web Development Easily</h2>
-        <p className="text-lg mb-4">
+        <h2 className="text-4xl text-gray-600 font-bold mb-6">Welcome to StackVarsity</h2>
+        <p className="text-lg text-gray-400 mb-4">
           Join our platform to learn, practice, and master web development skills.
         </p>
-        <Link href="/dashboard" className="bg-blue-600 text-white px-4 py-2 rounded hover:bg-blue-700">
-          Go to Dashboard
-        </Link>
       </main>
     </div>
   );
